requirements.txt
numpy>=1.21.0
pandas>=1.3.0
python-dateutil>=2.8.0
Flask>=2.0.0
python-dotenv>=0.19.0